backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==23.0.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
boto3==1.35.80
python-multipart==0.0.20
aiofiles==24.1.0
python-dotenv==1.0.1
